requirements.txt
annotated-types
anyio
beautifulsoup4
certifi
charset-normalizer
click
contourpy
cycler
faiss-cpu
fastapi
filelock
filterpy
fonttools
fsspec
ftfy
gdown
h11
idna
Jinja2
joblib
kiwisolver
lap
loguru
MarkupSafe
matplotlib
mpmath
networkx
numpy
opencv-python
packaging
pandas
pillow
psutil
pybind11
pycairo
pydantic
pydantic_core
PyGObject
pyparsing
PySocks
python-dateutil
python-multipart
pytz
PyYAML
regex
requests
scikit-learn
scipy
setuptools
shapely
six
sniffio
soupsieve
starlette
sympy
threadpoolctl
torch
torchaudio
torchvision
tqdm
typing_extensions
tzdata
urllib3
uvicorn
wcwidth
wheel
yolov5_openvino_object_detector
